Added docstrings for owl classes + refactoring part 2
