+++ owlapy/class_expression/class_expression.py
@@ -1,3 +1,4 @@
+"""OWL Base Classes Expressions"""
 from abc import abstractmethod, ABCMeta
 from ..data_ranges import OWLPropertyRange
 from ..meta_classes import HasOperands
