+++ owlapy/has.py
@@ -1,4 +1,7 @@
+"""Has index"""
 from typing import Protocol, ClassVar
+
+
 class HasIndex(Protocol):
     """Interface for types with an index; this is used to group objects by type when sorting."""
     type_index: ClassVar[int]  #: index for this type. This is a sorting index for the types.
