+++ owlapy/meta_classes.py
@@ -1,3 +1,5 @@
+"""Meta classes for OWL objects."""
+
 # https://docs.python.org/3/reference/datamodel.html#metaclasses
 from typing import TypeVar, Generic, Iterable
 from abc import ABCMeta, abstractmethod
