--- owlapy/owl2sparql/__init__.py
@@ -1,1 +1,0 @@
-"""OWL-to-SPARQL converter."""
