+++ owlapy/owl_individual.py
@@ -1,14 +1,22 @@
+"""OWL Individuals"""
 from abc import ABCMeta
 from .owlobject import OWLObject, OWLEntity
 from .iri import IRI
 from typing import Final, Union
+
+
 class OWLIndividual(OWLObject, metaclass=ABCMeta):
     """Represents a named or anonymous individual."""
     __slots__ = ()
     pass
 
+
 class OWLNamedIndividual(OWLIndividual, OWLEntity):
-    """Represents a Named Individual in the OWL 2 Specification."""
+    """Named individuals are identified using an IRI. Since they are given an IRI, named individuals are entities.
+        IRIs from the reserved vocabulary must not be used to identify named individuals in an OWL 2 DL ontology.
+
+        (https://www.w3.org/TR/owl2-syntax/#Named_Individuals)
+        """
     __slots__ = '_iri'
     type_index: Final = 1005
