+++ owlapy/owl_literal.py
@@ -1,7 +1,8 @@
+"""OWL Literals"""
 from abc import ABCMeta, abstractmethod
 from functools import total_ordering
 from .owl_annotation import OWLAnnotationValue
-from typing import Final, Optional, Union, TypeVar, Set
+from typing import Final, Optional, Union, Set
 from .types import OWLDatatype
 from datetime import datetime, date
 from pandas import Timedelta
@@ -9,10 +10,15 @@ from owlapy.vocab import OWLRDFVocabulary, XSDVocabulary
 from .owl_property import OWLObjectProperty, OWLDataProperty
 
 Literals = Union['OWLLiteral', int, float, bool, Timedelta, datetime, date, str]  #:
-_R = TypeVar('_R', bound='OWLPropertyRange')  #:
+
 
 class OWLLiteral(OWLAnnotationValue, metaclass=ABCMeta):
-    """Represents a Literal in the OWL 2 Specification."""
+    """Literals represent data values such as particular strings or integers. They are analogous to typed RDF
+    literals and can also be understood as individuals denoting
+    data values. Each literal consists of a lexical form, which is a string, and a datatype.
+
+     (https://www.w3.org/TR/owl2-syntax/#Literals)
+     """
     __slots__ = ()
 
     type_index: Final = 4008
@@ -255,6 +261,8 @@ class _OWLLiteralImplInteger(OWLLiteral):
     def get_datatype(self) -> OWLDatatype:
         # documented in parent
         return IntegerOWLDatatype
+
+
 class _OWLLiteralImplBoolean(OWLLiteral):
     __slots__ = '_v'
 
@@ -479,6 +487,7 @@ class _OWLLiteralImpl(OWLLiteral):
 
     def __repr__(self):
         return f'OWLLiteral({repr(self._v)}, {self._datatype})'
+
 
 #: the built in top object property
 OWLTopObjectProperty: Final = OWLObjectProperty(OWLRDFVocabulary.OWL_TOP_OBJECT_PROPERTY.get_iri())
