+++ owlapy/providers.py
@@ -8,32 +8,32 @@ from pandas import Timedelta
 Restriction_Literals = Union[OWLLiteral, int, float, Timedelta, datetime, date]
 
 
-def OWLDatatypeMaxExclusiveRestriction(max_: Restriction_Literals) -> OWLDatatypeRestriction:
+def owl_datatype_max_exclusive_restriction(max_: Restriction_Literals) -> OWLDatatypeRestriction:
     """Create a max exclusive restriction."""
     r = OWLFacetRestriction(OWLFacet.MAX_EXCLUSIVE, max_)
     return OWLDatatypeRestriction(r.get_facet_value().get_datatype(), r)
 
 
-def OWLDatatypeMinExclusiveRestriction(min_: Restriction_Literals) -> OWLDatatypeRestriction:
+def owl_datatype_min_exclusive_restriction(min_: Restriction_Literals) -> OWLDatatypeRestriction:
     """Create a min exclusive restriction."""
     r = OWLFacetRestriction(OWLFacet.MIN_EXCLUSIVE, min_)
     return OWLDatatypeRestriction(r.get_facet_value().get_datatype(), r)
 
 
-def OWLDatatypeMaxInclusiveRestriction(max_: Restriction_Literals) -> OWLDatatypeRestriction:
+def owl_datatype_max_inclusive_restriction(max_: Restriction_Literals) -> OWLDatatypeRestriction:
     """Create a max inclusive restriction."""
     r = OWLFacetRestriction(OWLFacet.MAX_INCLUSIVE, max_)
     return OWLDatatypeRestriction(r.get_facet_value().get_datatype(), r)
 
 
-def OWLDatatypeMinInclusiveRestriction(min_: Restriction_Literals) -> OWLDatatypeRestriction:
+def owl_datatype_min_inclusive_restriction(min_: Restriction_Literals) -> OWLDatatypeRestriction:
     """Create a min inclusive restriction."""
     r = OWLFacetRestriction(OWLFacet.MIN_INCLUSIVE, min_)
     return OWLDatatypeRestriction(r.get_facet_value().get_datatype(), r)
 
 
-def OWLDatatypeMinMaxExclusiveRestriction(min_: Restriction_Literals,
-                                          max_: Restriction_Literals) -> OWLDatatypeRestriction:
+def owl_datatype_min_max_exclusive_restriction(min_: Restriction_Literals,
+                                               max_: Restriction_Literals) -> OWLDatatypeRestriction:
     """Create a min-max exclusive restriction."""
     if isinstance(min_, float) and isinstance(max_, int):
         max_ = float(max_)
@@ -47,8 +47,8 @@ def OWLDatatypeMinMaxExclusiveRestriction(min_: Restriction_Literals,
     return OWLDatatypeRestriction(r_min.get_facet_value().get_datatype(), restrictions)
 
 
-def OWLDatatypeMinMaxInclusiveRestriction(min_: Restriction_Literals,
-                                          max_: Restriction_Literals) -> OWLDatatypeRestriction:
+def owl_datatype_min_max_inclusive_restriction(min_: Restriction_Literals,
+                                               max_: Restriction_Literals) -> OWLDatatypeRestriction:
     """Create a min-max inclusive restriction."""
     if isinstance(min_, float) and isinstance(max_, int):
         max_ = float(max_)
