+++ tests/test_class_expression_semantics.py
@@ -1,12 +1,7 @@
-from owlapy.iri import IRI
-
 from owlapy.class_expression import OWLClass, OWLObjectComplementOf, OWLObjectUnionOf
-from owlapy.class_expression import OWLBooleanClassExpression, OWLObjectIntersectionOf, OWLClassExpression
+from owlapy.class_expression import OWLObjectIntersectionOf, OWLClassExpression
 from owlapy.owl_property import OWLObjectProperty
 from owlapy.class_expression import OWLObjectSomeValuesFrom, OWLObjectAllValuesFrom
-
-from owlapy.owl2sparql.converter import owl_expression_to_sparql
-from owlapy.render import owl_expression_to_dl
 
 
 class TestClassExpression:
