+++ tests/test_owlapy_render.py
@@ -1,15 +1,18 @@
 import unittest
-from owlapy.owl_property import OWLObjectProperty
-from owlapy.model import OWLDataMinCardinality, OWLObjectIntersectionOf, OWLObjectSomeValuesFrom, \
-    OWLThing, OWLObjectComplementOf, OWLObjectUnionOf, OWLNamedIndividual, OWLObjectMinCardinality, IRI, OWLDataProperty, DoubleOWLDatatype, OWLClass, \
-    IntegerOWLDatatype, OWLDataExactCardinality, OWLDataHasValue, OWLDataAllValuesFrom, \
-    OWLDataOneOf, OWLDataSomeValuesFrom, OWLLiteral, BooleanOWLDatatype, \
-    OWLDataMaxCardinality
+
+from owlapy.iri import IRI
+from owlapy.owl_individual import OWLNamedIndividual
+from owlapy.owl_literal import DoubleOWLDatatype, IntegerOWLDatatype, OWLLiteral, BooleanOWLDatatype
+from owlapy.owl_property import OWLObjectProperty, OWLDataProperty
+from owlapy.class_expression import OWLDataMinCardinality, OWLObjectIntersectionOf, OWLObjectSomeValuesFrom, \
+    OWLThing, OWLObjectComplementOf, OWLObjectUnionOf, OWLObjectMinCardinality, OWLClass, \
+    OWLDataExactCardinality, OWLDataHasValue, OWLDataAllValuesFrom, \
+    OWLDataOneOf, OWLDataSomeValuesFrom, \
+    OWLDataMaxCardinality, OWLObjectHasValue, OWLObjectOneOf
 
 from owlapy.data_ranges import OWLDataComplementOf, OWLDataIntersectionOf, OWLDataUnionOf
-from owlapy.model.providers import OWLDatatypeMinMaxInclusiveRestriction
+from owlapy.providers import owl_datatype_min_max_inclusive_restriction
 from owlapy.render import DLSyntaxObjectRenderer, ManchesterOWLSyntaxOWLObjectRenderer
-from owlapy.class_expression import OWLObjectHasValue, OWLObjectOneOf
 
 class Owlapy_DLRenderer_Test(unittest.TestCase):
     def test_ce_render(self):
@@ -63,7 +66,7 @@ class Owlapy_DLRenderer_Test(unittest.TestCase):
         print(r)
         self.assertEqual(r, "∃ hasAge.¬xsd:double")
 
-        datatype_restriction = OWLDatatypeMinMaxInclusiveRestriction(40, 80)
+        datatype_restriction = owl_datatype_min_max_inclusive_restriction(40, 80)
 
         dr = OWLDataAllValuesFrom(property=has_age, filler=OWLDataUnionOf([datatype_restriction, IntegerOWLDatatype]))
         r = renderer.render(dr)
@@ -149,7 +152,7 @@ class Owlapy_ManchesterRenderer_Test(unittest.TestCase):
         print(r)
         self.assertEqual(r, "hasAge some not xsd:double")
 
-        datatype_restriction = OWLDatatypeMinMaxInclusiveRestriction(40, 80)
+        datatype_restriction = owl_datatype_min_max_inclusive_restriction(40, 80)
 
         dr = OWLDataAllValuesFrom(property=has_age, filler=OWLDataUnionOf([datatype_restriction, IntegerOWLDatatype]))
         r = renderer.render(dr)
